Style : Rename model file
