+++ event-model.js
